Fix: With require_event_comment set, the status of a new event with preexisting comment can be changed
Precondition: A new event was just commented, but the status flag was left at "new". Issue occurred with require_event_comment being true and when attempting to just change the status of that event (and not sending a new_comment to the backend). The backend expects new_comment to be set with each status change (except for NEW). Thus, the frontend was updated to set new_comment with every update.