Refactoring StatsService
Stats are no longer included in State data returned by the API. Consequently, state_year, state_month and stats_division are no longer sent to the API when fetching updates. That simplifies the periodic update process and didn't work for the dashboard anyway due to a regression from an earlier bug fix.