Fix: Re-render system overview panel periodically
The change event only fires when the state returned by the server differs from the local model. We always want to recalculate, because the final model state also depends on other models such as the sensor collection. Binding to the 'sync' instead of the 'change' event causes the handler to be executed on every sync attempt.